requirements.txt
torch>=2.0.0
sentence-transformers>=5.0.0
transformers>=4.50.0
peft>=0.18.0
huggingface-hub>=0.20.0
requests>=2.28.0
timm>=0.9.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
